python/virtualenv: Fix activate script to halt on error
The old version of the activate script would carry on trying to
install things if, e.g., virtualenv failed because it's not available.
We now explicitly test the result of each command (with `&&`) and do
not continue if anything fails.

This can't be done with `set -e` because this is code run in an
existing shell which may have that enabled and where we don't want to
change it.