python/virtualenv: Add (imperfect) raw download link for `activate`
Unfortunately, there's no way (at least that I can see) to get the raw
version with a relative link. So instead I link to the raw version
under `https://github.com/0cjs/sedoc/raw/master` which should usually
be the one you want anyway.

We use the link from the button on the web page, not the subsequent
link to `raw.githubusercontent.com` from the redirect. The redirect is
an HTTP 302 "Moved Temporarily," indicating that you should continue
to use the original URI for future requests. [RFC 2616]

[RFC 2616]: https://tools.ietf.org/html/rfc2616#section-10.3.3