python/runtime/activate: Show project name in prompt
This used to add `(virtualenv)` as the prefix to the prompt because
that's the subdirectory under `.build/` in which the virtualenv is
installed. Showing the base directory name gives a lot more
information, since then it's easier to see what project you're working
on. (People still know it's a virtualenv by the look of the prompt.)