python/activate: Bootstrap virtualenv in case we don't have it
Some systems may not have `virtualenv` installed in the path. We rid
ourselves of this dependency by doing a local "bootstrap" install of
virtualenv under `.build/virtualenv-bootstrap` and using that.

We still depend on having `pip` available, but we now try `python3 -m
pip` and if that doesn't work for whatever reason we fall back to
`python -m pip`. Thus the virtualenv defaults to the Python version
for which we have `pip` available, but this can of course be
overridden with the `--python` option to the activate script.

The Debian 9 `python-pip` package sets `--user` by default when
installing as a non-root user, which disallows the `--build` and
`--target` options; we explicitly set `--system` to undo this.
However, this will fail on non-Debian systems (with return code 2). We
need to figure out the best way to deal with this.