python/concurrency: Tweak for slightly better ease of use
Somehow a parenthetical comment about processes didn't stand out in
the README table of contents, so remove the parens.

Though the call() link at the top of the legacy API list also led to a
point near the docs for the following to functions, I keep missing
that so just add links for the following two as well.