ee/7400: Change to table of 5V logic levels
This is easier to read than the bullet-point version.

Also add a link to Sparkfun tutorial on logic levels. It says that
TTL max. Vout is 0.4 V, not 0.5 V. Interestingly, it has the exact
same diagrams as Radical Brad, but with that one voltage level
difference. I wonder where they come from?