python/runtime/activate: Avoid failure when -u is set
`set -u` will generate an error when unset variables are referenced,
with a few exceptions. $1 is not one of these exceptions, so this
script fails when called with -u set and no parameters.

We fix this by instead referencing array $@ with no subscript, which
returns the first element of the array, which is the same  as $1.
References $@ elements do not trigger an error even when -u is set.