python/runtime/activate: Previous $@ thing was very broken; fix
I don't know what I was smoking with the last commit, but $@ is not an
array; it's specially evaulated and always gives all the params. The
test case I didn't try was when passing two arguments to activate.

The proper way to deal with potentially undefined variables in the
face of `set -u` is to parameter expansion with a default value,
in this case `${1-}` to substitute an empty string if $1 is unset.