ascii.txt: Rework with 32-char cols, etc.
Using full 32-character columns (still with a blank line after
the 16th char) makes it much easier to read across from a control
code to see its corresponding letter.

Also, move the decimal value to the right of the character and
add vertical lines separating the columns, making it easier to
search down through the hex or decimal codes as necessary.

This also allows us to add indications at the right for groups
with bit 4 and bit 3 set.