python/activate: Update to handle virtualenv module layout change
The virtualenv package has changed the file layout for its modules,
moving `virtualenv.py` to `virtualenv/__init__.py` etc. So rather than
trying to follow filename changes, we now put the bootstrap directory
in $PYTHONPATH and use the -m option with the module name, letting
Python deal with whatever the filenames are.

This should be backward compatible with old versions of virtualenv.