8bit/apple2/monitor: Rewrite and clarify
This now abandons the "last opened location" and "next changable
location" terminology, as well as a lot else from the manual, because
the manual seems to be using an incorrect conceptual model of how the
monitor handles addresses and commands. (Maybe in part because the way
it does so is odd and a bit complex.) This new model I came up with
should be more accurate.