8bit/apple2/{dos,disk}: Link to HQ copy of _The DOS Manual_
Use archive.org/stream/ links to the "stream" viewer. It saves having
to manualy hit the "full page" button, which is what people usually
immmediately do, and at the top has an easy link back to the item
details and download

We also split the individual page references out to their own sections
in the reference list to make the overall reference list easier to
read and understand.