ee/ram-rom: Use ¶ to mark programming voltage pins
Some pins have an additional function activated by a higher voltage,
such as 14 V on 27512 O̅E̅ and A9 for erase or 9 V on 28C64 A9 for
identification area access.

The pilcrow ¶ is a handy way to quickly and visibly identify pins
supporting this general concept. Also, it's shorter than fully
spelling it out which will help if we try to narrow the columns.