ee/scope: Delete, move info to tools/Rigol-DS1000Z
ee/scope.md had only one tiny hint in it, so it's not worth a separate
file, and anyway I misss this file when looking up 'scope stuff. So
just move the hint to a generic section at the top of the Rigol file,
the only other one that deals with 'scopes.