docker/image: Slight rewrite that also fixes a formatting problem
GitHub normally has no trouble with back-quoted code spans that cross
line breaks. However, in this case the text after the break started
with a shell redirection `>` which for some reason it decided meant
that this was the start of a Markdown block quote.

I had to rewrite the sentence anyway; that shifted the line break
which, while still breaking the code span across two lines, no longer
starts with a `>` after the line break. GitHub renders this correctly.