ee/memory: Rework 24/28-pin table by size, rather than type
This works a lot better. It makes it easy to see the differences between
23xx/27xx/28Cxx/RAM for replacing one with another, and also just seems
to be a bit clearer in terms of the overall evolution of the packaging.

Now we can remove the '●' that was indicating RAM/ROM differences for
a pin since RAM and ROM for a given size are right next to each other.