linux/distro/pkg-nix: Change NixOS wiki links to new wiki site
The previous wiki on nixos.wiki was hosted by a third party which was
unresponsive to various problems so, as of early 2024, the wiki has been
forked to wiki.nixos.org, which is self-hosted. (The old wiki site remains,
however.) Further details are in the FAQ[1] under "Why is there a new wiki?
What is with nixos.wiki"

[1]: https://wiki.nixos.org/wiki/FAQ

Reviewed-by: Curt J. Sampson <cjs@cynic.net>