nix/: Moved from linux/distro/pkg-nix
We move our Nix package system documentation to the top level because
a) it's not actually Linux-only, because it supports MacOS too, and
b) it's expanding enough that we need to split it into multiple files.

The old file is left behind with just a link to nix/README.md. That
contains most of the old information, except what's been extracted to
nix/errors.md and nix/install.md. Some further Arch Linux information
has been added to nix/install.md.

Much of the information in nix/README.md needs to be extracted to
separate files.