linux/distro/bld-debian: Proper sections for binary and source packages
Turns out that the _Debian Administrator's Handbook,_ which does not come
up anywhere near first in searches, has the concise and accurate
information I am looking for, unlike the half-assed crap that comes up at
the top of searches. Now I have good sections for this, good leads on
what to read next, and I need to clean up the remainder of the file
based on this new information.